Use DeveloperLevel rather that Developer Mode. Re-use Extra Data field to pass level.

+++ Meadow.CLI.Core/Devices/IMeadowDevice.cs
@@ -39,7 +39,7 @@ namespace Meadow.CLI.Core.Devices
         public Task NshDisable(CancellationToken cancellationToken = default);
         public Task TraceEnable(CancellationToken cancellationToken = default);
         public Task SetTraceLevel(uint traceLevel, CancellationToken cancellationToken = default);
-        public Task SetDeveloper(uint mode, uint userData, CancellationToken cancellationToken = default);
+        public Task SetDeveloper(ushort mode, uint userData, CancellationToken cancellationToken = default);
         public Task SetDeveloper1(uint userData, CancellationToken cancellationToken = default);
         public Task SetDeveloper2(uint userData, CancellationToken cancellationToken = default);
         public Task SetDeveloper3(uint userData, CancellationToken cancellationToken = default);
